Move script to @wordpress/scripts
