+++ packages/scripts/scripts/packages-update.js
@@ -1,31 +1,24 @@
 /* eslint-disable no-console */
-
 /**
  * External dependencies
  */
-const chalk = require( 'chalk' );
-
-/**
- * Internal dependencies
- */
-const { readJSONFile, runShellScript, askForConfirmationToContinue } = require( './utils' );
+const fs = require( 'fs' );
+const spawn = require( 'cross-spawn' );
 
 /**
  * Constants
  */
 const WORDPRESS_PACKAGES_PREFIX = '@wordpress/';
-const BUILD_COMMAND = 'npm run build';
+
+function readJSONFile( fileName ) {
+	const data = fs.readFileSync( fileName, 'utf8' );
+	return JSON.parse( data );
+}
 
 function getWordPressPackages( packageJSON ) {
 	return Object.keys( packageJSON.dependencies ).
 		concat( Object.keys( packageJSON.devDependencies ) ).
 		filter( ( packageName ) => ( packageName.startsWith( WORDPRESS_PACKAGES_PREFIX ) ) );
-}
-
-function getCommandUpdatePackagesLatestVersion( packages ) {
-	const packagesWithLatest = packages.map( ( packageName ) => ( `${ packageName }@latest` ) );
-	const packagesWithLatestJoined = packagesWithLatest.join( ' ' );
-	return `npm install ${ packagesWithLatestJoined } --save`;
 }
 
 function getPackageVersionDiff( initialPackageJSON, finalPackageJSON ) {
@@ -48,6 +41,17 @@ function getPackageVersionDiff( initialPackageJSON, finalPackageJSON ) {
 	return diff.sort( ( a, b ) => a.dependency.localeCompare( b.dependency ) );
 }
 
+function updatePackagesToLatestVersion( packages ) {
+	const packagesWithLatest = packages.map(
+		( packageName ) => ( `${ packageName }@latest` )
+	);
+	return spawn.sync( 'npm', [
+		'install',
+		...packagesWithLatest,
+		'--save',
+	], { stdio: 'inherit' } );
+}
+
 function outputPackageDiffReport( packageDiff ) {
 	console.log( [
 		'The following package versions were changed:',
@@ -60,41 +64,11 @@ function outputPackageDiffReport( packageDiff ) {
 function updatePackageJSON() {
 	const initialPackageJSON = readJSONFile( 'package.json' );
 	const packages = getWordPressPackages( initialPackageJSON );
-	const packageUpdateCommand = getCommandUpdatePackagesLatestVersion( packages );
-	runShellScript( packageUpdateCommand );
+	const result = updatePackagesToLatestVersion( packages );
 	const finalPackageJSON = readJSONFile( 'package.json' );
 	outputPackageDiffReport( getPackageVersionDiff( initialPackageJSON, finalPackageJSON ) );
+	process.exit( result.status );
 }
 
-async function updateCorePackages() {
-	const abortMessage = 'Aborting!';
-	console.log(
-		chalk.bold( '💃 Time to update WordPress core packages 🕺\n\n' ),
-		'Welcome! This tool is going to help you with updating core WordPress packages to their latest version.\n',
-		'The working directory should be set to the WordPress develop.\n'
-	);
-	await askForConfirmationToContinue(
-		'Proceed with the update and install of the latest version for @wordpress packages?',
-		true,
-		abortMessage
-	);
-	updatePackageJSON();
-
-	await askForConfirmationToContinue(
-		'Run the build so remaining files (e.g: php block changes) are updated?',
-		true,
-		abortMessage
-	);
-	runShellScript( BUILD_COMMAND );
-
-	console.log(
-		'\n>> 🎉 The changes were applied to the WordPress directory.\n',
-		'Please test and submit a patch with these changes.\n',
-		'Thank you for updating the core packages!'
-	);
-}
-
-module.exports = {
-	updateCorePackages,
-};
+updatePackageJSON();
 /* eslint-enable no-console */
